implementacion principios solid

+++ src/main/resources/db/migration/V3__create_table_espacios_fisicos.sql
@@ -2,7 +2,7 @@ create table espacios_fisicos(
     id bigint not null auto_increment,
     id_tipo_espacio bigint not null,
     nombre varchar(100) not null unique,
-    capacidad tinyint not null,
+    capacidad smallint not null,
     descripcion varchar(200),
 
     primary key (id),
